Increase delay before minetest aborts when SQLite3 database is busy
The current timeout of 3 seconds can sometimes be exceeded when
minetest is scanning a very large world. The new timeout of 10 seconds
means significant lag, *if* it happens, but that's still preferable
to the alternative of aborting minetest.
A timeout much larger than 10 seconds seems a bad idea, as that will
cause players to lose too much work if minetest ends up aborting after
all.