Fix interference between malloc and zstd
This interference causes a severe performance degradation:

On every compression, zstd will allocate and free memory.
It turns out that each time, this causes malloc to request a new
5MB mmap()ped area from the OS, which is zeroed using memset. The
subsequent free() will cause that 5MB area to be munmap()ped again.

This repeated and unnecessary allocating, zeroing and freeing of
a mmapped area consumes countless unnecessary CPU cycles.

This patch tells malloc not to use mmap(), and if necessary, also
to only release a significant amount of memory to the OS.