Move map limits verification to MapgenParams
The map generation limits are mapgen parameters.
Move them to MapgenParams, and implement all verification of
positions against boundaries there as well.

This fixes possible bugs as well, as previously, the configured
maximum was used in computations. However, due to the way map
generation works (blocks, chunks, etc.), with a chunk size of 5,
that maximum can be up to 95 nodes larger than the actual maximum.